Move new_task, switch_mode, attempt_completion, and ask_followup_question to tool files (#2102)
