+++ src/core/tools/types.ts
@@ -13,3 +13,7 @@ export type HandleError = (action: string, error: Error) => Promise<void>
 export type PushToolResult = (content: ToolResponse) => void
 
 export type RemoveClosingTag = (tag: ToolParamName, content?: string) => string
+
+export type AskFinishSubTaskApproval = () => Promise<boolean>
+
+export type ToolDescription = () => string
